Fix Regex Error
The `ops` object gets joined and put between square brackets (ex: [÷×+-]). The problem is `-` is used to identify ranges and when placed between characters (ex: [÷×-+]) rather than at the start or end of a character class (ex: [÷×+-]) it will either produce an error or be treated as the range indicator and not as the literal `-` character. The fix is simply to ensure the `-` is at the start or end of this object structure.